fix: change parameter type of wrap_int from int to long long, improve typecasting in all wrapper functions

+++ src/expr.h
@@ -125,7 +125,7 @@ int is_null_expr(Expr_t);
 
 void print_expr(Expr_t);
 
-Expr_t wrap_int(int);
+Expr_t wrap_int(long long);
 Expr_t wrap_flt(double);
 Expr_t wrap_str(char*);
 Expr_t wrap_bool(char);
